2DAW/Python/Actividades_Faltantes/requirements.txt
pandas
openpyxl
numpy
flask
matplotlib
scikit-learn
pygame
mysql-connector-python
yfinance
